+++ src/TestSuite/TestCase.php
@@ -15,8 +15,7 @@ namespace DatabaseBackup\TestSuite;
 
 use Cake\Core\Configure;
 use Cake\TestSuite\TestCase as CakeTestCase;
-use DatabaseBackup\TestSuite\TestCaseTrait;
-use Tools\TestSuite\TestCaseTrait as ToolsTestCaseTrait;
+use Tools\TestSuite\TestCaseTrait;
 
 /**
  * TestCase class
@@ -24,7 +23,6 @@ use Tools\TestSuite\TestCaseTrait as ToolsTestCaseTrait;
 abstract class TestCase extends CakeTestCase
 {
     use TestCaseTrait;
-    use ToolsTestCaseTrait;
 
     /**
      * Setup the test case, backup the static object values so they can be
@@ -36,6 +34,7 @@ abstract class TestCase extends CakeTestCase
     {
         parent::setUp();
 
+        Configure::write('DatabaseBackup.connection', 'test');
         $this->loadPlugins(['DatabaseBackup']);
     }
 
@@ -48,7 +47,48 @@ abstract class TestCase extends CakeTestCase
     {
         parent::tearDown();
 
-        Configure::write('DatabaseBackup.connection', 'test');
         $this->deleteAllBackups();
+    }
+
+    /**
+     * Internal method to create a backup file
+     * @return string
+     */
+    protected function createBackup()
+    {
+        return $this->BackupExport->filename('backup.sql')->export();
+    }
+
+    /**
+     * Internal method to creates some backup files
+     * @param bool $sleep If `true`, waits a second for each backup
+     * @return array
+     */
+    protected function createSomeBackups($sleep = false)
+    {
+        $files[] = $this->BackupExport->filename('backup.sql')->export();
+
+        if ($sleep) {
+            sleep(1);
+        }
+
+        $files[] = $this->BackupExport->filename('backup.sql.bz2')->export();
+
+        if ($sleep) {
+            sleep(1);
+        }
+
+        $files[] = $this->BackupExport->filename('backup.sql.gz')->export();
+
+        return $files;
+    }
+
+    /**
+     * Internal method to deletes all backups
+     * @return void
+     */
+    public function deleteAllBackups()
+    {
+        safe_unlink_recursive(Configure::read('DatabaseBackup.target'));
     }
 }