+++ tests/TestCase/Utility/BackupImportTest.php
@@ -89,7 +89,7 @@ class BackupImportTest extends TestCase
     /**
      * Test for `filename()` method, with invalid directory
      * @expectedException ErrorException
-     * @expectedExceptionMessageRegExp /^File or directory `[\s\w\/:\\]+backup\.sql` is not readable$/
+     * @expectedExceptionMessageRegExp /^File or directory `[\s\w\/:\\\-]+backup\.sql` is not readable$/
      * @test
      */
     public function testFilenameWithInvalidDirectory()